Added EMBED_DEL_OBSOLETE.BAT.
This script deletes files from previous Embed software installations
that should no longer be present in new installations.  For example,
if a new program is created that replaces a previous BAT file, the old
BAT file must be deleted from new installations that include the new
program.

The intent is to keep this script up to date as files are obsoleted.